refactor(sync): inline room manager into server-local
sync-core is now pure protocol primitives (encode/decode + handlers).
rooms.ts moved to server-local where it's exclusively used. Collapsed
WsSyncPluginConfig duplication, removed broken ./discovery export.

- Completed spec items: all waves (1, 2, 3)
- Added y-protocols as direct dep to server-local
- Verified @epicenter/sync actively uses sync-core (kept)
