+++ packages/server-local/package.json
@@ -20,7 +20,8 @@
 		"@elysiajs/openapi": "^1.4.11",
 		"elysia": "^1.2.25",
 		"typebox": "catalog:",
-		"wellcrafted": "catalog:"
+		"wellcrafted": "catalog:",
+		"y-protocols": "catalog:"
 	},
 	"peerDependencies": {
 		"@epicenter/workspace": "workspace:*",
