+++ packages/sync-core/package.json
@@ -4,8 +4,7 @@
 	"main": "./src/index.ts",
 	"types": "./src/index.ts",
 	"exports": {
-		".": "./src/index.ts",
-		"./discovery": "./src/discovery/index.ts"
+		".": "./src/index.ts"
 	},
 	"license": "AGPL-3.0",
 	"scripts": {
